Consolidate some Dockerfile commands to create less layers.
We have to COPY over a ton of files for compile assets to work.
Each COPY command creates a new docker layer for the image.
This is a quick fix to consolidate some of those into single COPY
commands so less layers are created. The inspiration for this was
troubleshooting why this container wasn't working in Heroku and
there is this mention "Images with more than 40 layers may fail to start in the Common Runtime"
here: https://devcenter.heroku.com/articles/container-registry-and-runtime#known-issues-and-limitations

The proper solution to making a smaller image and reducing overhead is prob
to split our build up into a multi-stage build where the compile time and dev
only stuff is separate from the prod image.
See: https://docs.docker.com/develop/develop-images/multistage-build/