More tinkering with heroku.yml for troubleshooting.
Still can't see app logs and still crashing. I'm guessing it crashes b/c
the app isn't binding to the port within 60 second, but i need to see logs
to know what the heck is going on. So I'm adding some more echo statments to
see if the ENV vars that are supposed to configure the logs are properly coming through.