Try to reap leaked database connections.
Seems like we are leaking one connection per worker process.
E.g. if there are 5 threads and the pool is 5, there should be
a max of 5 connections. I went into the database, cleared all the
connection for this app, restarted the app server and watched what
happened re: connections and it always ends up with 6 per worker.
I'm going to use the ENV['DB_POOL'] and ENV['RAILS_MAX_THREADS']
to have the pool be 1 larger than the threads, but this commit
is a separate attempt to reclaim that 1 leaked connection b/c
who knows, there may end up being more once this server is actually
under load. It could be thread's dying without closing the connection
or something.

See: https://api.rubyonrails.org/v4.0.13/classes/ActiveRecord/ConnectionAdapters/ConnectionPool.html

This is a pretty detailed article and about this stuff:
https://bibwild.wordpress.com/2014/07/17/activerecord-concurrency-in-rails4-avoid-leaked-connections/